Remove old internal module.
